feat: first pass at browser bundles for all libs
