+++ packages/core/package.json
@@ -14,7 +14,7 @@
     "prepublish": "npm run build",
     "prebuild": "npm run clean",
     "build": "npm run compile",
-    "build:webpack": "webpack --mode production",
+    "build:browser": "webpack --mode production --config ../../webpack.config.js",
     "compile": "tsc -b tsconfig.json",
     "clean": "rimraf ./dist ./tsconfig.tsbuildinfo",
     "test:browser": "polendina --cleanup --timeout 60 src/**/*.spec.ts --webpack-config ../../webpack.test.js"
