--- tsconfig.webpack.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "compilerOptions": {
-    "sourceMap": true
-  },
-  "exclude": [
-    "**/*.spec.ts",
-    "node_modules"
-  ]
-}
